% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCtNormalize.R
\name{deltaCtNormalize}
\alias{deltaCtNormalize}
\title{the function outputs a matrix of normalized Ct values according to a house keeping gene}
\usage{
deltaCtNormalize(mat, hk = "RPLP0")
}
\arguments{
\item{mat}{is a matrix with sample ids as rownames and genes as column names}

\item{hk}{is the name of the house keeping gene}
}
\description{
For a matrix of TLDA or RT-PCR expression it will normalize the Ct values for running linear models
}
