% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_top.R
\name{extract_top}
\alias{extract_top}
\title{creates an output table including all genes in ebfit}
\usage{
extract_top(
  BigTab = NULL,
  ebfit,
  pcut = 0.05,
  cont = NULL,
  method = "none",
  n = 75,
  direction = "both",
  coefColumn
)
}
\arguments{
\item{BigTab}{is a big tab object with GENENAME as the column with name of the gene}

\item{ebfit}{it is an object from eBayes function in limma package}

\item{pcut}{this is the cutoff for significance}

\item{cont}{this is the label of the contrast that is prefixed with lgFCH, FCH, pvals , fdr or status}

\item{method}{a parameter indicating which method is used for type-I error control (none, fdr , bonferroni)}

\item{n}{the number of genes to be included in the list of top differentially expressed}

\item{direction}{the direction of regulation induced by the treatment (both = both directions, up = upregulation, down = downregulation)}
}
\description{
For a pre-specified contrast this function will return the top differentially expressed genes
}
