% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.Zscores.R
\name{generate.Zscores}
\alias{generate.Zscores}
\title{The function outputs a data frame in long format with rownames as sample ids, geneset names and zscores}
\usage{
generate.Zscores(mat, geneset, geneset.name)
}
\arguments{
\item{mat}{is a matrix of expressions with genes as rownames and samples ids as columns}

\item{geneset}{is a vector with gene symbols}

\item{geneset.name}{is a string with the name of the gene set}
}
\description{
A function that transform values into Z scores
}
