% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{minimal}
\alias{minimal}
\title{This function creates a minimal example dataset
The output of this function is a matrix of log-fold changes and a matrix of p-values}
\usage{
minimal()
}
\description{
Generation of artificial dataset for testing purposes
}
