% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerize.R
\name{numerize}
\alias{numerize}
\title{the function output a numeric vector that represents numbers previously represented as strings}
\usage{
numerize(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
the function convert numbers that are strings to numbers
}
