% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval2star.R
\name{pval2star}
\alias{pval2star}
\title{The function outputs stars associated with the level of significance of a p-value}
\usage{
pval2star(
  x,
  bk = c(-Inf, 0.001, 0.01, 0.05, 0.1, Inf),
  lb = c("***", "**", "*", "+", "")
)
}
\arguments{
\item{x}{is a vector of p-values}

\item{bk}{is a vector with cutoffs to establish significance levels}

\item{lb}{is a vector with labels for the significance levels}
}
\description{
For a vector of p-values, this function will output a set of star values
}
