% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCor.R
\name{runCor}
\alias{runCor}
\title{The function outputs a vector from a data frame or a tibble}
\usage{
runCor(varx, vary, dat = newdata)
}
\arguments{
\item{data}{a data frame of tibble}

\item{var}{the name of the column to be extracted}
}
\description{
The function extract a column from a data frame and converts it into a vector
}
